# public/openapi.yaml
openapi: 3.0.3
info:
  title: Croatian Legal AI War Machine API
  description: |
    Comprehensive API for Croatian legal system analysis, prosecutorial misconduct detection,
    and AI-powered legal research. This system provides advanced capabilities for:

    - Legal document search and retrieval (laws, court decisions, cases)
    - OpenAI API proxy with Croatian legal context
    - Prosecutorial misconduct detection and legal action generation
    - Multi-agent collaboration for complex legal problems
    - Predictive analytics for case outcomes
    - Graph-based citation analysis
    - MCP (Model Context Protocol) tools integration

    ## Authentication

    Most endpoints require API token authentication via the `Authorization: Bearer <token>` header.
    Some endpoints use `X-MCP-Token` header for MCP-specific authentication.

    ## Rate Limiting

    - Standard endpoints: 60 requests per minute
    - MCP endpoints: Tool-specific limits
    - Info endpoints: 60 requests per minute

    ## Croatian Legal Context

    This API is specifically designed for the Croatian legal system and includes:
    - Croatian law codes (KZ, ZKP, ZSZD, Ustav RH)
    - Croatian court hierarchy (Općinski sud, Županijski sud, Vrhovni sud)
    - Croatian legal procedures (žalba, zaštita zakonitosti, ustavna tužba)
  version: 1.0.0
  contact:
    name: API Support
    email: support@legal-ai.example.com
  license:
    name: Proprietary
    url: https://legal-ai.example.com/license
servers:
...
security:
  - BearerAuth: []
  - McpToken: []
paths:
...
  /search:
    post:
      tags: [Search]
      summary: Unified search
      description: Search across all legal corpora (laws, decisions, cases)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnifiedSearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedSearchResponse'
  /search/laws:
    post:
      tags: [Search]
      summary: Search laws
      description: Search Croatian laws only
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Law search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /search/decisions:
    post:
      tags: [Search]
      summary: Search court decisions
      description: Search court decisions only
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Decision search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /search/cases:
    post:
      tags: [Search]
      summary: Search private cases
      description: Search user's private cases
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Case search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /search/hybrid:
    post:
      tags: [Search]
      summary: Hybrid search
      description: Combines vector and keyword search for better results
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/SearchRequest'
                - type: object
                  properties:
                    vector_weight:
                      type: number
                      default: 0.7
                    keyword_weight:
                      type: number
                      default: 0.3
      responses:
        '200':
          description: Hybrid search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /search/with-citations:
    post:
      tags: [Search]
      summary: Search with citation context
      description: Search that includes citing and cited documents
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/SearchRequest'
                - type: object
                  properties:
                    citation_depth:
                      type: integer
                      default: 2
      responses:
        '200':
          description: Search results with citations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitationSearchResponse'

  # Reasoning Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API token authentication via Authorization header

    McpToken:
      type: apiKey
      in: header
      name: X-MCP-Token
      description: MCP-specific authentication token
...
  schemas:
...
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        limit:
          type: integer
          default: 10
        filters:
          type: object
        include_embeddings:
          type: boolean
          default: false
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              score:
                type: number
              content:
                type: string
              metadata:
                type: object
        total:
          type: integer
        query_time:
          type: number
    UnifiedSearchRequest:
      allOf:
        - $ref: '#/components/schemas/SearchRequest'
        - type: object
          properties:
            corpora:
              type: array
              items:
                type: string
                enum: [laws, decisions, cases]
              default: [laws, decisions]
    UnifiedSearchResponse:
      type: object
      properties:
        laws:
          type: array
        decisions:
          type: array
        cases:
          type: array
        total:
          type: integer
    CitationSearchResponse:
      allOf:
        - $ref: '#/components/schemas/SearchResponse'
        - type: object
          properties:
            citations:
              type: object
              properties:
                citing:
                  type: array
                cited_by:
                  type: array

    # MCP Tool Schemas
...